move server routes to a submodule
