+++ mithril-aggregator/src/http_server/mod.rs
@@ -1,8 +1,4 @@
-mod certificate_routes;
-mod middlewares;
+mod routes;
 mod server;
-mod signatures_routes;
-mod signer_routes;
-mod snapshot_routes;
 
 pub use server::{Server, SERVER_BASE_PATH};
