+++ mithril-aggregator/src/http_server/routes/mod.rs
@@ -1,0 +1,6 @@
+mod certificate_routes;
+mod middlewares;
+pub mod router;
+mod signatures_routes;
+mod signer_routes;
+mod snapshot_routes;
